Sockets Controles de permisos según tipo de Usuario

+++ src/persona/persona.controller.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Body, Controller, Delete, Get, NotFoundException, Param, Patch, Post, Res, ValidationPipe} from '@nestjs/common';
+import { BadRequestException, Body, Controller, Delete, Get, Logger, NotFoundException, Param, Patch, Post, Res, Session, ValidationPipe} from '@nestjs/common';
 import { PersonaCreateDto } from './dto/persona.create.dto';
 import { Persona } from './persona.entity';
 import { PersonaService } from './persona.service';
@@ -6,42 +6,90 @@ import { validate } from 'class-validator';
 import { resolveMx, resolveNaptr } from 'dns';
 import { PersonaUpdateDto } from './dto/persona.update.dto';
 import { throws } from 'assert';
+import { RolService } from 'src/rol/rol.service';
 
 @Controller('persona')
 export class PersonaController{
-    constructor(private readonly _personaServices:PersonaService){}
+    constructor(
+        private readonly _personaServices:PersonaService,
+        private readonly _rolServices:RolService){}
+
+    private logger:Logger = new Logger('PersonaController');
 
     @Get('all')
     async buscarPersonas(
-        @Res() res
+        @Res() res,
+        @Session() session
     ){
-        const personas = await this._personaServices.findAll();
-        res.send({results:personas});
-        return personas;
+        try {
+            if(await this._rolServices.isUserType(session,['Admin','Empleado'])){
+                res.status(403).send({
+                  "statusCode": 403,
+                  "message": "Forbidden resource",
+                  "error": "Forbidden"
+                });
+                return;
+            }
+
+            const personas = await this._personaServices.findAll();
+            res.send({results:personas});
+            return personas;
+        } catch (error) {
+            this.logger.error(error);
+        }
     }
 
     @Get(':idPersona')
     async buscarPersonaID(
         @Res() res,
-        @Param('idPersona') idPersona:string
+        @Param('idPersona') idPersona:string,
+        @Session() session
     ){
-        const personaEncontrada = await this._personaServices.findByID(idPersona);
-        res.send({personaEncontrada:personaEncontrada});
+        try {
+            if(await this._rolServices.isUserType(session,[])){
+                res.status(403).send({
+                  "statusCode": 403,
+                  "message": "Forbidden resource",
+                  "error": "Forbidden"
+                });
+                return;
+            }
+            const personaEncontrada = await this._personaServices.findByID(idPersona);
+            res.send({personaEncontrada:personaEncontrada});
+        } catch (error) {
+            this.logger.error(error);
+        }
     }
 
     @Delete('del/:idPersona')
     async eliminarPersonaID(
         @Res() res,
-        @Param('idPersona') idPersona:string
+        @Param('idPersona') idPersona:string,
+        @Session() session
     ){
-        const eliminado = await this._personaServices.deletePerson(idPersona);
-        res.send({eliminado})
+        try {
+            if(await this._rolServices.isUserType(session,['Admin','Empleado'])){
+                res.status(403).send({
+                  "statusCode": 403,
+                  "message": "Forbidden resource",
+                  "error": "Forbidden"
+                });
+                return;
+            }
+            const eliminado = await this._personaServices.deletePerson(idPersona);
+            this.logger.log(`Persona ${idPersona} eliminada.`)
+            this.logger.log(`Usuario resposable: ${session.usuario.u_username} - ${session.rol.r_rol}`);
+            res.send({eliminado})
+        } catch (error) {
+            this.logger.error(error);
+        }
     }
 
     @Patch('update/:idPersona')
     async actualizarPersonaID(
         @Res() res,
         @Body() persona:Persona,
+        @Session() session,
         @Param('idPersona') idPersona?
     ){
         delete persona['p_cedula'];
@@ -51,6 +99,14 @@ export class PersonaController{
         person.p_apellidos = persona.p_apellidos;
         
         try {
+            if(await this._rolServices.isUserType(session,[])){
+                res.status(403).send({
+                  "statusCode": 403,
+                  "message": "Forbidden resource",
+                  "error": "Forbidden"
+                });
+                return;
+            }
             const errores = await validate(person);
             if(errores.length>0){
                 console.error(errores);
@@ -66,7 +122,7 @@ export class PersonaController{
                 res.status(200).send({personaActualizada});
             }
         } catch (error) {
-            
+            this.logger.error(error);
         }
 
     }
@@ -74,13 +130,22 @@ export class PersonaController{
     @Post('crear')
     async crearPersona(
         @Res() res,
-        @Body() persona:Persona
+        @Body() persona:Persona,
+        @Session() session
     ){
         const person = new PersonaCreateDto();
         person.p_nombres = persona.p_nombres;
         person.p_apellidos = persona.p_apellidos;
         person.p_cedula = persona.p_cedula;
         try {
+            if(await this._rolServices.isUserType(session,['Admin','Empleado'])){
+                res.status(403).send({
+                  "statusCode": 403,
+                  "message": "Forbidden resource",
+                  "error": "Forbidden"
+                });
+                return;
+            }
             const errores = await validate(person);
 
             if(errores.length>0){
@@ -91,7 +156,7 @@ export class PersonaController{
                 res.send({personaCreada:personaCreada})
             }
         } catch (error) {
-            console.error(error);
+            this.logger.error(error);
         }
     }
